(#9583) Fix provider detection for gentoo and unsupported linuxes for the iptables provider.
Previously we had fairly specific confine settings for named distributions
of linux for the iptables provider. This was silly, since the commands
defined in the provider should be enough to confine the provider to Linux
only systems.

I've removed the confine, and replaced the defaultfor to be:

  :kernel => :linux

Which should avoid the need to keep adding extra Linux distributions.

I've also added some spec tests for provider detection which should help
catch any failures around the command based detection in the future.