(#10026) Add aliases for test,specs,tests to Rakefile and provide -T as default.
I've cleaned up the Rakefile a bit so that the aliases:

* test
* specs
* tests

Now point to "spec". The (s) versions are used by rvm but generally these are
all provided for convenience in case someone uses the wrong form.

I've also changed the default rake task to show the list of tasks (rake -T).

The description for rake spec is also a bit longer.