(#10162) add firewallchain type and iptables_chain provider
Add firewallchain type and iptables_chain provider. This is required to
support the firewall class and it is invisiged that an autorequire will
be used to automaticly create user chain. This type can also set
policies on inbuilt chains.

Currently this is only a prototype. It contains a number of issues
needing direction as described on
http://projects.puppetlabs.com/issues/10162.