(#14755) Stub iptables_version for now so tests run on non-Linux hosts
Without a stub some tests fail on non-Linux hosts. This is because they are
expecting a particular version of iptables to exist which isn't always true.

The right answer for the provider is to actually allow the fact to be set
per test, but for now we are doing a global override just to make tests pass.