Replace Execution.stubs with provider.expects
Regression introduced in PR #102.

This wasn't satisfying the call to `iptables-save` and was allowing the test
to fallback to the real implementation. On OSX machines where the command
doesn't exist it resulted in the failure:

     Puppet::ExecutionFailure:
       Execution of '/sbin/iptables-save' returned 1:

It's likely that it passed on Travis CI because the command was available.

Using expects instead of stubs here because we know that it should be
called and want to supersede any `before :each` condition.