Always stub iptables-save command
Regression introduced in PR #102. Three tests were failing when run on OSX:

- iptables provider should be able to get a list of existing rules
- iptables provider when creating resources insert_args should be an array
- iptables provider when modifying resources update_args should be an array

This is because they were attempting to access the real binary
`iptables-save` which doesn't exist resulting in:

     Puppet::ExecutionFailure:
       Execution of '/sbin/iptables-save' returned 1:

In order to ensure that we never hit the real binary we always stub
`iptables-save` with empty content before each test. Unless that specific
test expects content, then it can use expects like the previous commit.