Fix severe installation hang on jessie.
On jessie iptables-persistent pops up debconf to ask if the current firewall setup should be saved which leads to puppet hanging and 99% of the time half configured firewall and ssh lockout.

This patch sets the value to false before the install so that debconf doesnt pop up - iptables rules shouldn't be saved automatically since they will be overridden by puppet'.