add -g flag handling in ip6tables.rb provider
In lib/puppet/provider/firewall/ip6tables.rb there is no goto: entry for the -g flag in @resource_map which leads to errors for rules using that feature. 

lib/puppet/type/firewall.rb already handles this flag. It's a one-line change to recognize it. The right rules then get into the chains in my testing.