(maint) Typo fixes (#844)
* Changed case of error message

Changed the case of the tables in the error message when defining a custom chain. The error message shows the tables in uppercase (as you'd see when using iptables directly), while the regex in lib/puppet/provider/firewallchain/iptables_chain.rb is expecting lowercase. It looks like lib/puppet/provider/firewallchain/iptables_chain.rb is expecting the default chains in uppercase so I have left those.

* Corrected typo, hyphen not yphen