(CONT-5) Raising minimum required puppet version
Prior to this commit, and as part of the codebase hardening project, we
sanitised some commands in the module. However, this sanitisation
method requires that users have at least puppet version 6.24.0 and
onwards to be effective.

This commit aims to resolve that requirement by raising the required
puppet version in the metadata.json file.