Add multi-version support to provider unit tests
Add support for testing against multiple kernel and iptables versions,
to ensure that features gated by kernel or iptables version aren't
rendered un-testable.  Also, pull the handy construct for handling the
addition of '--wait' in later iptables versions in the ip6tables test
cases over to iptables tests as well, because with multi-version
support, the test cases for iptables will start to need it as well.