Add cases exercised in acceptance to unit testing
Add the test cases not present in unit testing, but used in acceptance
testing, to the unit tests as well.  Not everyone who works on the
module has the ability to run the acceptance tests, but acceptance test
failures should not come as a surprise to developers being diligent
about passing all unit tests.