Group target match options right behind targets
The @resource_list determines the order in which arguments appear on the
generated commandline for ip[6]tables; if there are options which belong
to a match module interleaved with options which belong to a target
extension, and both the target and match extensions are used in the same
rule, perverse outputs like '-j SYNPROXY -m tcpmss --mss 1360
--sack-perm' are possible - where an entire match extension and its
arguments are inserted between a target extension and its arguments.
Group options which belong to a target extension together immediately
behind :jump, so a target extension spec and its arguments cannot be
separated; and add comments in the array to ensure it remains clear
why this is important.