Teach the parser context-awareness
Not all arguments are globally unique across all iptables extensions,
and we should not search the entire rule string for all arguments when
each one is only ever valid within a known context.  Define start and
stop delimiters for existing known overlaps, and default to a
full-string search if no start or end markers are defined.  Also, stop
being quite so terse; the Ruby authors do not charge by the letter for
variable names, and readability is good.