Remove trash from the middle of the main parse
The regular parse mechanisms can handle table, chain, and comment(s) on
their own already, if a single space character is added to the head of
the rule.  :table was already in @resource_map, :chain is added to the
parse output by the unnecessary special mechanism anyway, and :name
already grabs ALL comments successfully since the escaping fixes.
'line' is used instead of 'line.dup' because string addition returns a
new string, so String#dup would be redundant.