Reverse the parse direction
Parse the rule from the head to prevent the parser from finding
aliases of the intended argument in the rule's heead.

String#slice! searches from the head of a string, so if the parser is
using it to iteratively search for strings ordered progressively further
from the back of a rule, each iteration is a new opportunity to find an
aliased substring in its head and completely moot the point of
generating an ordered parse list in the first place.

Additionally:
- Untie the logical knots that were required for reverse-parsing
- Don't use Enumerable#each to do absolutely everything by side effect
  when an iterator could just return the constructed value directly
- Don't pre-define variables we don't need yet just so they can be
  reassigned
- Freshen up pointlessly terse variable names that only exist in an
  ephemeral scope anyway, this is already hard enough to follow
- Stop our value regex from capturing a second, nested group if we only
  need the outer to avoid any later need to transpose the result
- Fix any remaining variables with misleading names (e.g., valrev is no
  longer "values, reversed") or errant comments that no longer
  accurately describe the operation of the code