Make mode options for '-m recent' parse correctly
Options setting the mode of operation for the 'recent' match extension
no longer parse from existing rule definitions with their long-option
double-dashes still included.  Note that these options are meant to be
bang-invertible, but this module does not support that use case (and
currently will explode if it encounters any such rule).