Fix managing build-in chain on iptables nft
`iptables-save` does not show any build-in chains when they have not
been interacted with. This leads to the chain being (re)-created.

PR#1206 attempted to fix this, by listing the chain instead of creating
it, but this method only seems to work on non-nft iptables[1].

Fix this for nft version of iptables by setting the policy of the chain
instead of listing it. This seems to work for both nft and non-nft
version of iptables (tested on Ubuntu 20.04 and 24.04).

Fixes #1217

[1] https://github.com/puppetlabs/puppetlabs-firewall/issues/1217#issuecomment-2130211647